# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal branding website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  timestamp: string;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  techStack: string[];
  githubUrl?: string | null;
  liveUrl?: string | null;
  imageUrl?: string | null;
  featured: boolean;
  order: number;
  createdAt: string;
}

export interface ContactFormInput {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  email: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  subject: string;
  /**
   * @minLength 10
   * @maxLength 5000
   */
  message: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  subject: string;
  message: string;
  createdAt: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminAuthResult {
  message: string;
  username: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}