# lib/api-zod/src/generated/types/adminAuthResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal branding website API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminAuthResult {
  message: string;
  username: string;
}