# lib/api-zod/src/generated/types/contactFormInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal branding website API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactFormInput {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  email: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  subject: string;
  /**
   * @minLength 10
   * @maxLength 5000
   */
  message: string;
}