# lib/api-zod/src/generated/types/contactSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal branding website API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  subject: string;
  message: string;
  createdAt: Date;
}